レッスン画面をサブコンポーネントへ分離
