+++ src/components/IndexPage/LessonPageContent/IdolInformation.tsx
