+++ src/components/IndexPage/LessonPageContent/ScoreInformation.tsx
@@ -1,6 +1,6 @@
 import { Idol, IdolParameterKind } from "gakumas-core";
 import React from "react";
-import { idolParameterKindToTextColorClassName } from "./utils";
+import { idolParameterKindToTextColorClassName } from "../utils";
 
 type Props = {
   idolParameterKind: IdolParameterKind;
