+++ src/components/IndexPage/LessonPageContent/SkipButton.tsx
